+++ packages/editor/types.ts
@@ -2,5 +2,6 @@ export type CustomText = { text: string };
 
 export type CustomElement = {
   type: string;
+  collapsed?: boolean;
   children: (CustomElement | CustomText)[];
 };